+++ tests/dispatch_cascade.c
@@ -39,7 +39,7 @@ union {
 	char padding[64];
 } indices[BLOCKS];
 
-uint32_t iterations = (size_t)(QUEUES * BLOCKS * 0.25);
+uintptr_t iterations = (size_t)(QUEUES * BLOCKS * 0.25);
 
 void
 histogram(void) {
@@ -79,7 +79,7 @@ histogram(void) {
 
 void
 cascade(void* context) {
-	uint32_t idx, *idxptr = (uint32_t*)context;
+	uintptr_t idx, *idxptr = (uintptr_t*)context;
 
 	if (done) return;
